requirements.txt
annotated-types==0.5.0
blinker==1.6.2
click==8.1.5
dnspython==2.3.0
Flask==2.3.2
Flask-Cors==4.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.0.3
pydantic_core==2.3.0
pymongo==4.4.1
python-dotenv==1.0.0
typing_extensions==4.7.1
Werkzeug==2.3.6
zipp==3.16.2